util.poll: Increase max epoll events per call
This makes the struct roughly the same size in both epoll and select
mode (64bit).

There doesn’t seem to be much guidance on an appropriate size, it does
not seem to matter too much since if there are more events they will
simply show up in the next epoll_wait call. The number of exactly
concurrent events should be fairly low most of the time anyways.