mod_pep: Remove use of recipients table for signaling pending disco#items
It stored the caps hash as a string in the recipients table while
waiting for a disco#info response, then replaces it with the set of
nodes that had +notify.

This mixing of types is bad, so here it gets removed.

This can be improved once an IQ tracking API (#714) is added.