MUC: Don't advertise registration feature on host JID (fixes #1451)
There is currently no mention in XEP-0045 of how or where to advertise
support for registration.

Advertising on the host JID may be confusable with service-wide
registration, as implemented in ejabberd.

A common and sensible pattern in XMPP is that a feature is advertised on
the JID where the service is available.