mod_admin_telnet: Pretty-print values returned from commands
This makes it much nicer to inspect Prosody internals.

Existing textual status messages from commands are not serialized to
preserve existing behavior. Explicit serialization of configuration is
kept in order to make it clear that returned strings are serialized
strings that would look like what's actually in the config file.

The default maxdepth of 2 seems ought to be an okay default, balanced
between showing enough structure to continue exploring and DoS-ing your
terminal.

Thanks to Ge0rG for the motivation to finally do this.