mod_external_services: Validate services added via events
While writing developer documentation it became obvious that i was silly
to have one item format for config and items API, and another format for
the event API.

Then there's the stanza format, but that's a common pattern.

This change reduces the possible input formats to two and allows other
modules the benefit of the processing and validation performed on items
from the config.