mod_storage_internal: Use a binary search for time based ranges
Iterating over an entire archive to find a few items in the far end from
where iteration started is expensive, and probably more expensive with
the lazy-loading of items added in the previous commit.

Since we can now efficiently read items in random order, we can now use
a binary search to find a better starting point for iteration.