mod_s2s: Bail if connection is destroyed after attempting to open stream
Fixes "attempt to compare number with nil" because `session.version` has
been cleared by s2smanager.destroy_session.

This can happen with the server_epoll setting opportunistic_writes
enabled, which means that it can notice that the connection failed at
this point, after which it triggers the whole chain of events that leads
to session destruction and "cleaning" most of the session fields.