mod_bosh,mod_websocket: Make into global-shared modules (...again)
Global modules aren't quite considered loaded onto hosts, which
causes confusion in some cases. They are also reported in the log as
being served on http://*:5280/foo which is also a bit confusing, and
can't be clicked.

Global modules also have to have their paths configured in the global
section, which could be confusing and unexpected.

This global+shared method should be the best of both worlds.