mod_s2s: Guard against missing 'to' on incoming stream
Given an incoming <stream:stream from="example.com"> this line would
have mistakenly reported the 'from' as the local host. Neither are
technically required and may be missing, especially on connections used
only for Dialback.

Outgoing connections initiated by Prosody always have 'from_host' and
'to_host', so it is safer to check it this way.