util.pubsub: Fix behavior of persist_items disabled
When set to 'false' there is no need for a persistence interface at all,
since items are not persisted after being broadcast.

Had started wondering if maybe the behavior was wrong, after reading
parts of XEP-0060 that pointed in that direction.

Some discussion of this can be found in logs of
xmpp:xsf@muc.xmpp.org?join from around 2021-07-20

Thanks to Ralph for confirming.