makefile: fix prosody.version target
POSIX is quite explicit regarding the precedence of AND-OR lists [0]:

> The operators "&&" and "||" shall have equal precedence and shall be
> evaluated with left associativity. For example, both of the following
> commands write solely `bar` to standard output:
>       false && echo foo || echo bar
>       true || echo foo && echo bar

Given that, `prosody.version` target behaves as

	((((((test -f prosody.release && cp ...) ||
	    test -f ...) &&
	    sed ...) ||
	    test -f ...) &&
	    hexdump ...) ||
	    echo unknown > $@)

In the case of release tarballs, `prosody.release` does exist, so the
first AND pair is executed. Given that it's successful, then the first
`test -f` in the OR pair is ignored, and instead the `sed` in the AND
pair is executed. `sed` success, as `.hg_archival.txt` exists, making
the second `test -f` in the OR pair ignored, and `hexdump` in the AND
pair is executed. Now, given that `.hg` doesn't exist, it fails, so the
last `echo` is run, overwriting `prosody.version` with `unknown`.

This can be worked around placing `()` around the AND pairs. Decided to use
conditionals instead, as I think they better communicate the intention
of the block.

[0]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_09_03