mod_pubsub: Move reversal of item order in <items>
Examples in XEP-0060 suggest that items should be listed in
chronological order, but we get them from the archive in reverse
order.

However when requesting specific items by id the results keep that
order and we don't want to flip it again.

At some point it would likely be best to use the archive API directly
instead of this util.cache-compatible wrapper.