net.server_epoll: Try to make port number related methods sane
Previously it was unclear whether "client port" was the port that the
client connected to, or from. I hereby declare that the client port is
the source port and the server port is the destination port.

Incoming and outgoing connections can be distinguished by looking at
the_server reference, which only incoming connections have.