mod_s2s: Fix logging of <stream:error> consistency with other mods
`reason` was often a table, so the log said "table: 0xptr" or such.
mod_c2s, mod_bosh etc logs the stream error stanza object, so better do
the same.  It would be nicer if this was an util.error object, but that
will have to be a future change.