util.prosodyctl.check: Move word to ease future translations
Recent experience with translations in the context of Snikket
highlighted that sentences spread across concatenated strings like this
makes the experience less than pleasant for translators.

We don't have translation yet, but it is a future goal and why not?

The duplication can be solved with a parameterized function for the
common cases.