mod_http_file_share: Always measure total disk usage for statistics!
Metrics available or not depending on configuration is weird, even tho
it might be expensive to calculate and it's only really needed when
there is a global quota.

Default quota is set to infinity, which is essentially what it was.

Reports NaN if there is an error, which should count as over the
infinite default quota.