mod_cron: Allow for a small amount of timer drift
If the timer activates a bit early then a task might be just a few
seconds short of being allowed to run. This would run such a task rather
than wait another hour.

The value 0.5% chosen so that a weekly task does not run an entire hour
earlier than last time.