util.xml: Break reference to help the GC (fix #1711)
LuaExpat uses a registry reference to track handlers, which makes
it so that an upvalue like this creates a reference loop that keeps the
parser and its handlers from being garbage collected. The same issue has
affected util.xmppstream in the past.

Code for checking:

	local xml_parse = require"util.xml".parse;
	for i = 1, 10000 do xml_parse("<root/>") end
	collectgarbage(); collectgarbage();
	print(collectgarbage("count"), "KiB");

A future release of LuaExpat may fix the underlying issue there.