mod_s2s: Retrieve TLS context for outgoing Direct TLS connections from mod_tls
So that the same TLS context is used for both Direct TLS and starttls,
since they are supposed to be functionally identical apart from the few
extra round trips.

A new event is added because the 's2s-created' event fires much later,
after a connection has already been established, where we need the TLS
context before that.