modulemanager: Add plugin load filter that reads module metadata from source
Metadata in modules is added using lines formatted as:

--% key: value

Where key is a valid identifier string, and value is also a string (leading
and trailing whitespace are trimmed during parsing).

The initial supported keys are:

--% requires_core_features: feature1, feature2, ...
--% conflicts_core_features: feature1, feature2. ...

These 'features' map to features reported by the new core.features module.

A benefit of this load-time metadata approach compared to e.g. something like
module:requires()/module:conflicts() is that we can continue to look in module
search paths for a suitable module. Aborting an already-loaded module due to
a version conflict would be too late.