prosody.cfg.lua.dist: Remove comment about mod_*.lua above modules_enabled
This is a very old statement, but people generally don't need to check for the
files, and shouldn't be encouraged to put them in Prosody's source dir. The
installer will be the way forward for most people, and hg for the rest.

Manually moving files into the right place is not something most users should
be doing.